refactor: separate the code into reusable components
